src/requirements.txt
streamlit
openai
sqlalchemy-iris
